+++ exoline/plugins/spec.py
@@ -38,7 +38,7 @@ import sys
 import pyonep
 import ast
 
-import yaml
+import ruamel.yaml as yaml
 import jsonschema
 import requests
 import six
@@ -650,6 +650,7 @@ datarules:
                             desc = myinfo['description']
                             is_script = desc['format'] == 'string' and 'rule' in desc and 'script' in desc['rule']
                             if is_script:
+                                os.makedirs(script_dir)
                                 filename = os.path.join(script_dir, info['aliases'][rid][0])
                                 spec.setdefault('scripts', []).append({'file': filename})
                                 with open(filename, 'w') as f: